test: moved mocksdirectory to graphql/__mocks__
