+++ lib/graphql/resolvers/mutations/__tests__/addSkill.spec.ts
@@ -1,5 +1,5 @@
 import { addSkill } from '../addSkill'
-import { ctx } from '../../__mocks__/apolloServerContext'
+import { ctx } from '../../../__mocks__/apolloServerContext'
 
 const args = {
   skill: {
