+++ lib/graphql/resolvers/mutations/__tests__/login.spec.ts
@@ -1,6 +1,6 @@
 import { login } from '../login'
 import { defaultRequest } from '../../../../services/consents'
-import { ctx } from '../../__mocks__/apolloServerContext'
+import { ctx } from '../../../__mocks__/apolloServerContext'
 
 jest.mock('../../../../services/consents', () => ({
   defaultRequest: jest.fn(() => 'defaultRequest'),
